Increasing number of batches when single precision isn't large enough to store the batch size * N (for adjacency graph & CSR row in array.
Will eventually expose an argument through Python to allow the user to control the precision of the outputs (and thus the precision of the internal structures.). It's important to allow the user to set this explicitly because there's a huge tradeoff to be made between a larger number of much faster batches and a smaller number of much slower batches.

+++ cpp/src/dbscan/vertexdeg/naive.h
@@ -38,7 +38,7 @@ static const int TPB_Y = 8;
  * @param N number of rows
  * @param D number of columns
  */
-template <typename Type, typename Index_ = long>
+template <typename Type, typename Index_ = int>
 __global__ void vertex_degree_kernel(Pack<Type, Index_> data,
                                      Index_ startVertexId, Index_ batchSize) {
   const Type Zero = (Type)0;
